lightweight_engine/extensions/hires_fix.py
"""
HiRes.fix на основе анализа логов
"""
import os
from typing import List, Dict, Any
from PIL import Image

class HiResFix:
    """
    HiRes.fix для улучшения разрешения как в логах
    """
    
    def __init__(self, config):
        self.config = config
        self.upscalers = {}
        self._initialize_upscalers()
        
    def _initialize_upscalers(self):
        """Инициализация upscalers как в логах"""
        for name, filename in self.config.UPSCALERS.items():
            upscaler_path = os.path.join(self.config.UPSCALERS_PATH, filename)
            if os.path.exists(upscaler_path):
                size_mb = os.path.getsize(upscaler_path) / 1024 / 1024
                print(f"[Setup] ✓ {name}: {upscaler_path} ({size_mb:.1f} MB)")
                self.upscalers[name] = upscaler_path
            else:
                print(f"[Setup] ✗ {name}: {upscaler_path} не найден")
                
    def process(self, images: List[Image.Image], scale: float, 
                steps: int, upscaler: str) -> List[Image.Image]:
        """Обработка HiRes.fix как в логах"""
        if scale <= 1.0:
            return images
            
        print(f"Starting HiRes.fix: scale={scale}, steps={steps}, upscaler={upscaler}")
        
        results = []
        for i, image in enumerate(images):
            # Upscaling как в логах
            upscaled_image = self._upscale_image(image, scale, upscaler)
            
            # Refinement как в логах (steps шагов)
            refined_image = self._refine_image(upscaled_image, steps)
            
            results.append(refined_image)
            
        return results
        
    def _upscale_image(self, image: Image.Image, scale: float, upscaler: str) -> Image.Image:
        """Upscaling изображения"""
        if upscaler not in self.upscalers:
            print(f"Warning: Upscaler {upscaler} not found, using simple resize")
            new_size = (int(image.width * scale), int(image.height * scale))
            return image.resize(new_size, Image.LANCZOS)
            
        print(f"Using {upscaler} for upscaling {scale}x")
        
        # Здесь будет реальная загрузка и использование ESRGAN/UltraSharp
        # Пока используем простое увеличение
        new_size = (int(image.width * scale), int(image.height * scale))
        upscaled = image.resize(new_size, Image.LANCZOS)
        
        return upscaled
        
    def _refine_image(self, image: Image.Image, steps: int) -> Image.Image:
        """Refinement изображения как в логах"""
        print(f"Running HiRes refinement: {steps} steps")
        
        # Здесь будет интеграция с основным движком для refinement
        # Симуляция прогресса как в логах
        for step in range(steps):
            progress = (step + 1) / steps * 100
            print(f"HiRes step {step + 1}/{steps} ({progress:.0f}%)")
            
        return image
        
    def get_available_upscalers(self) -> List[str]:
        """Получение списка доступных upscalers"""
        return list(self.upscalers.keys())