lightweight_engine/requirements.txt
# Основные (из логов)
torch>=2.6.0
torchvision
numpy
pillow

# Модели
safetensors
transformers>=4.0.0
diffusers>=0.20.0

# Оптимизации
accelerate
bitsandbytes==0.45.3

# ADetailer
ultralytics
opencv-python

# API
fastapi
uvicorn

# Утилиты
requests
tqdm